backend/requirements.txt
flask
mysql-connector-python
pyjwt
werkzeug
python-dotenv